Implement parsing of quic_transport_parameters extension
This should mostly implement support for the quic/ NPF. It needs c++
code to parse the quic_transport_parameters extension.

The NPF formatting currently is not 100% correct - this will probably
need a small rewrite of how the extension parameters are stored and
formatted.

There also currently is a small difference of how we display some of the
quic_transport_parameters; I have not yet identified the reason for
this.